feat: create QuantityControls component
