+++ src/pages/Cart/CartList/cartList.module.scss
@@ -106,72 +106,6 @@
         }
       }
     }
-
-    &__controls {
-      height: 4.4rem;
-      display: flex;
-      align-items: center;
-      gap: 1rem;
-      font-weight: 500;
-
-      @media all and (max-width: 480px) {
-        font-size: 1.8rem;
-      }
-
-      .controls__count,
-      .controls__buttons_btn {
-        box-shadow: none;
-        border-radius: 0;
-        height: inherit;
-      }
-
-      .controls__buttons {
-        height: inherit;
-        display: flex;
-        gap: 1px;
-
-        &_btn {
-          width: 4.4rem;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          flex-shrink: 0;
-          background-color: $dark;
-          background-size: 50%;
-          background-repeat: no-repeat;
-          background-position: center;
-
-          &:last-child {
-            background-image: url('../../../assets/icons/ic-plus.svg');
-          }
-
-          &:first-child {
-            background-image: url('../../../assets/icons/ic-minus.svg');
-          }
-
-          &:hover:not(:disabled) {
-            background-color: $primary;
-          }
-
-          &:disabled {
-            cursor: default;
-            opacity: 0.5;
-          }
-        }
-      }
-
-      .controls__count {
-        position: relative;
-        width: 6rem;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        border: 1px solid $dark;
-        text-align: center;
-        font-weight: 400;
-        user-select: none;
-      }
-    }
   }
 
   .footer {
