+++ src/pages/Cart/QuantityControls/index.ts
@@ -1,0 +1,2 @@
+import QuantityControls from './QuantityControls';
+export { QuantityControls };
